requirements.txt
scikit-learn
requests
beautifulsoup4
aiohttp
numpy
sentence-transformers
torch
pandas
python-dotenv
fastapi
uvicorn[standard]
DateTime
pyjwt
sqlalchemy
motor
pydantic
asyncpg
pymongo
bcrypt==4.0.1
passlib[bcrypt]
